# backend/src/main/resources/static/openapi/openapi.yaml
openapi: 3.0.1
info:
  title: LoLchievement API
  version: 1.0.0

servers:
  - url: http://localhost:8080

paths:
  ### Achievement
  /api/achievements/{pUUID}:
    get:
      operationId: getPlayerAchievementByPUUID
      summary: Get all player achievements by pUUID
      tags:
        - Achievement
      parameters:
        - name: pUUID
          in: path
          required: true
          schema:
            type: string
          description: Player uuid
      responses:
        '200':
          description: Player achievement information
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/AchievementDTO'
  ### Player
  /api/players/{id}:
    get:
      operationId: getPlayerById
      summary: Get player by PUUID
      tags:
        - Player
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
          description: Player uuid
      responses:
        '200':
          description: Player information
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PlayerDTO'
  /api/players/{gameName}/{tagName}:
    get:
      operationId: getPlayerByGameAndTagName
      summary: Get player by gameName and tagName
      tags:
        - Player
      parameters:
        - name: gameName
          in: path
          required: true
          schema:
            type: string
          description: In game player name
        - name: tagName
          in: path
          required: true
          schema:
            type: string
          description: Player tag name
      responses:
        '200':
          description: Player information
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PlayerDTO'

components:
  schemas:
    PlayerDTO:
      $ref: './components/schemas/playerDTO.yaml'
    AchievementDTO:
      $ref: './components/schemas/achievementDTO.yaml'